Added user authentication

+++ part4/blogApp/controllers/blogs.js
@@ -1,20 +1,15 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
-blogsRouter.get('/', async (request, response) => {
+blogsRouter.get('/', async (req, res) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
-  response.json(blogs)
+  res.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
+blogsRouter.post('/', middleware.userExtractor, async (req, res) => {
+  const body = req.body
+  const user = req.user
 
   const blog = new Blog({
     title: body.title,
@@ -28,23 +23,22 @@ blogsRouter.post('/', async (request, response) => {
     blog.likes = 0
   }
   if (!blog.title || !blog.url) {
-    response.status(400).send(blog)
+    res.status(400).send(blog)
   }
 
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
-  response.status(201).json(savedBlog)
-
-  //else {
-  //   blog.save().then(result => {
-  //     response.status(201).json(result)
-  //   })
-  // }
+  res.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (req, res) => {
+  const blog = await Blog.findById(req.params.id)
+  if (blog.user.toString() !== req.user.id) {
+    return res.status(401).json({ error: 'invalid user' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
